Un-nest `Common::Base` as just `Base`.
