+++ lib/fakturoid/api/expense.rb
@@ -3,7 +3,7 @@
 module Fakturoid
   module Api
     class Expense
-      include Common::Base
+      include Base
 
       def all(params = {})
         request_params = Utils.permit_params(params, :since, :updated_since, :page, :subject_id, :custom_id, :number, :variable_symbol, :status) || {}
