+++ lib/fakturoid/api/inbox_file.rb
@@ -3,7 +3,7 @@
 module Fakturoid
   module Api
     class InboxFile
-      include Common::Base
+      include Base
 
       def all(params = {})
         request_params = Utils.permit_params(params, :page)
